fix . 더보기 버튼 분리수정
