+++ public/js/play.js
@@ -4,7 +4,6 @@ import { HotSearchKeywordUI } from './search.hot.js';
 import { CarouselCtroller } from './carousel.ctrl.js';
 import { RequestData } from './request_data.js';
 import { URL } from './url.js';
-import { viewMoreUrlSetting } from './view_more_manager';
 
 const { CAROUSEL_API, VIEW_MORE_API, VIEW_MORE_API_CUSTOM } = URL;
 const requestInfoForSection1_carousel = 'mileageList';
@@ -40,6 +39,37 @@ const ctrlHotDealCarousel = () => {
     value: 'carousel_hot',
   };
   return new CarouselCtroller(carouselObj);
+};
+
+const resetViewMoreSection = () => {
+  let viewMores = _.$('.evt_list .panel');
+  while (viewMores.childNodes.length) {
+    viewMores.removeChild(viewMores.firstChild);
+  }
+
+  playViewMore();
+  count = 0;
+  changeButton(count);
+};
+
+const changeButton = (count) => {
+  let newBtn = count === 0 ? document.createTextNode('더보기') : document.createTextNode(' 접기');
+
+  const block = _.$('.see_more_btn');
+  block.removeChild(block.firstChild);
+  block.appendChild(newBtn);
+};
+
+const viewMoreUrlSetting = () => {
+  var viewMorePage = 2;
+  while (count < viewMorePage) {
+    count++;
+    let viewMoreUrl = VIEW_MORE_API_CUSTOM(count);
+    if (count === viewMorePage) changeButton(count);
+    return new RequestData(viewMoreUrl, requestInfoForSection2_view_more, 'view_more');
+  }
+
+  resetViewMoreSection();
 };
 
 const ctrlViewMoreBtn = () => {
